complete-chatbot-main/requirements.txt
langchain
langgraph
langchain-core
langchain-groq
langchain-google-genai
google-generativeai
python-dotenv